Move model_base.h from common/ to models/
