+++ src/examples/mnist/model.h
@@ -8,9 +8,9 @@
 
 #include "common/definitions.h"
 #include "common/keywords.h"
-#include "common/model_base.h"
 #include "graph/expression_graph.h"
 #include "layers/convolution.h"
+#include "models/model_base.h"
 
 #include "examples/mnist/dataset.h"
 #include "models/encdec.h"
