+++ src/rescorer/rescorer.h
@@ -3,14 +3,12 @@
 #include "marian.h"
 
 #include "common/config.h"
-#include "common/model_base.h"
 #include "data/batch_generator.h"
 #include "data/corpus.h"
 #include "graph/expression_graph.h"
-
-#include "models/model_task.h"
+#include "models/model_base.h"
 #include "models/model_factory.h"
-
+#include "models/model_task.h"
 #include "rescorer/score_collector.h"
 
 namespace marian {
