+++ src/training/graph_group.h
@@ -9,8 +9,8 @@
 
 #include "3rd_party/threadpool.h"
 #include "common/definitions.h"
-#include "common/model_base.h"
 #include "data/batch_generator.h"
+#include "models/model_base.h"
 #include "optimizers/optimizers.h"
 #include "training/dropper.h"
 #include "training/scheduler.h"
